drop `flutter_bloc` dependency
