+++ packages/widgetbook_core/lib/src/navigation_tree/navigation_tree.dart
@@ -1,4 +1,3 @@
-export 'bloc/navigation_bloc.dart';
 export 'enums/enums.dart';
 export 'models/models.dart';
 export 'widgets/widgets.dart';
