+++ packages/widgetbook_core/pubspec.yaml
@@ -14,11 +14,9 @@ dependencies:
   collection: ^1.15.0
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.1
   freezed_annotation: ^2.2.0
 
 dev_dependencies:
-  bloc_test: ^9.1.0
   build_runner: ^2.3.0
   flutter_test:
     sdk: flutter
