+++ widgetbook_for_widgetbook/lib/navigation/navigation_tree.dart
@@ -2,13 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:widgetbook_annotation/widgetbook_annotation.dart';
 import 'package:widgetbook_core/widgetbook_core.dart';
 
-import 'navigation_bloc_provider.dart';
 import 'navigation_test_data.dart';
 
 @UseCase(name: 'Default', type: NavigationTree)
 Widget navigationTreeDefaultUseCase(BuildContext context) {
-  return const NavigationBlocProvider(
+  return const NavigationTree(
     directories: directories,
-    child: NavigationTree(),
+    searchQuery: '',
   );
 }
