+++ widgetbook_for_widgetbook/pubspec.yaml
@@ -10,7 +10,6 @@ dependencies:
   cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.1
   widgetbook: ^3.0.0-beta.13
   widgetbook_annotation: ^3.0.0-beta.6
   widgetbook_core:
